refactor(validators): move update-user validator to auth validators directory
Add comprehensive validation for user update fields including email, password, phone, name and lastname
Include tests covering all validation scenarios and edge cases
